feat: addrequirements and make requested changes
