+++ gists/requirements-gists.txt
@@ -1,0 +1,3 @@
+pandas==2.1.3
+Requests==2.31.0
+ruamel.yaml==0.18.5
